requirements-test.txt
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-asyncio==0.21.1
pytest-xdist==3.3.1
pytest-timeout==2.1.0
pytest-selenium==4.0.1
pytest-env==1.0.1
freezegun==1.2.2
